requirements.txt
altgraph==0.16.1
future==0.16.0
keyboard==0.13.2
macholib==1.11
pefile==2018.8.8
Pillow==5.3.0
pkg-resources==0.0.0
pynput==1.4
python-xlib==0.23
six==1.11.0
